+++ main.rb
@@ -3,24 +3,7 @@ require_relative 'store'
 require_relative 'person'
 require_relative 'books'
 require_relative 'rentals'
-
-module InputHandler
-  def self.get_integer_input(prompt)
-    print "#{prompt}:"
-    gets.chomp.to_i
-  end
-
-  def self.get_string_input(prompt)
-    print "#{prompt}:"
-    gets.chomp.to_s
-  end
-
-  def self.get_boolean_input(prompt)
-    print "#{prompt}:"
-    answer = gets.chomp.downcase
-    answer == 'y'
-  end
-end
+require_relative 'input_Handler'
 
 module StartupMessage
   def self.print_once
